Se agrego correpcion el migracion routes

+++ database/migrations/2024_09_19_162814_create_routes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('origin'); // Origen de la ruta
             $table->string('destination'); // Destino de la ruta
             $table->float('distance'); // Distancia de la ruta
-            $table->decimal('price',15,2);
+            $table->integer('price');
             $table->enum('status', ['pendiente', 'en curso', 'entregada'])->default('pendiente'); // Estado de la entrega
 
             $table->timestamps();
